60. Using 3rd-Party Middleware ( morgan )
ref:  https://www.udemy.com/course/nodejs-express-mongodb-bootcamp/learn/lecture/15064872

also ref:  https://www.udemy.com/course/nodejs-express-mongodb-bootcamp/learn/lecture/15064868
59. Creating Our Own Middleware (pipeline mechanism in express.js)

Middleware的解說 ref:
https://github.com/avgsteve/natour-nodejs-project/commit/7105379f599f2f7cc7873c283c9d8be826e0dedb

=== >>>> 'morgan'   - HTTP request logger middleware for node.js
ref:  https://www.npmjs.com/package/morgan#dev

'morgan' ==> index.js
https://github.com/expressjs/morgan/blob/master/index.js

To have a look inside a middleware function ex:
module.exports = morgan
next()  etc,.

=== >>>>  Express.js Middleware resource page:
https://expressjs.com/en/guide/using-middleware.html

from the page:

Middleware functions can perform the following tasks:

Execute any code.
Make changes to the request and the response objects.
End the request-response cycle.
Call the next middleware function in the stack.

===>>> Another middleware used in this project:  app.use(express.json())

ref:  https://expressjs.com/en/4x/api.html#express.json
This is a built-in middleware function in Express. It parses incoming requests with JSON payloads and is based on body-parser.

Returns middleware that only parses JSON and only looks at requests where the Content-Type header matches the type option. This parser accepts any Unicode encoding of the body and supports automatic inflation of gzip and deflate encodings.

A new body object containing the parsed data is populated on the request object after the middleware (i.e. req.body), or an empty object ({}) if there was no body to parse, the Content-Type was not matched, or an error occurred.

As req.body’s shape is based on user-controlled input, all properties and values in this object are untrusted and should be validated before trusting. For example, req.body.foo.toString() may fail in multiple ways, for example foo may not be there or may not be a string, and toString may not be a function and instead a string or other user-input.