87. Refactoring for MVC (also setup mongoose schema and model's file for database)
ref:  https://www.udemy.com/course/nodejs-express-mongodb-bootcamp/learn/lecture/15065072

1. create models folder for MVC structure and tourModel.js

2.  create mongoose schema obj for making new data template and instance

Schema
ref: https://mongoosejs.com/docs/guide.html#models

Everything in Mongoose starts with a Schema. Each schema maps to a MongoDB collection and defines the shape of the documents within that collection.

(mongoose) Models:
ref: https://mongoosejs.com/docs/models.html

Models are fancy constructors compiled from Schema definitions. An instance of a model is called a document. Models are responsible for creating and reading documents from the underlying MongoDB database.