122. Catching Uncaught Exceptions ( process.on(uncaughtException,callback) )
ref:  https://www.udemy.com/course/nodejs-express-mongodb-bootcamp/learn/lecture/15065228

1. In server.js:

process.on('uncaughtException', err => {

ref:
https://nodejs.org/api/process.html#process_event_uncaughtexception

err <Error> The uncaught exception.

origin <string> Indicates if the exception originates from an unhandled rejection or from synchronous errors. Can either be 'uncaughtException' or 'unhandledRejection'.

The 'uncaughtException' event is emitted when an uncaught JavaScript exception bubbles all the way back to the event loop. By default, Node.js handles such exceptions by printing the stack trace to stderr and exiting with code 1, overriding any previously set process.exitCode. Adding a handler for the 'uncaughtException' event overrides this default behavior. Alternatively, change the process.exitCode in the 'uncaughtException' handler which will result in the process exiting with the provided exit code. Otherwise, in the presence of such handler the process will exit with 0.