168.(!!! important) Calculating Average Rating on Tours - Part 2
ref:  https://www.udemy.com/course/nodejs-express-mongodb-bootcamp/learn/lecture/15065554

1. In model Schema, create a static functions which can be used in Query obj later on

    ===>  reviewSchema.statics.calcAverageRatings = async function(tourId) {

Ref:
link: https://mongoosejs.com/docs/api.html#model_Model.aggregate

Mongoose does not cast aggregation pipelines to the model's schema because $project and $group operators allow redefining the "shape" of the documents at any stage of the pipeline, which may leave documents in an incompatible format. You can use the mongoose-cast-aggregation plugin to enable minimal casting for aggregation pipelines.
The documents returned are plain javascript objects, not mongoose documents (since any shape of document can be returned).

2. Use Query post middle ware to update the Tour data with lastest aggregation data from static method "calcAverageRatings" after the document is saved to db

  ===>  reviewSchema.post('save', function(next) {

3. Use .pre to get doc Query and pass it to next .post middl ware to update data when updating or deleting review data.

====>  reviewSchema.pre(/^findOneAnd/, async function(next) {

====>  reviewSchema.post(/^findOneAnd/, async function() {