169. Preventing Duplicate Reviews (by .index() in Schema)
ref:  https://www.udemy.com/course/nodejs-express-mongodb-bootcamp/learn/lecture/15065556

1.  Mongoose:  https://mongoosejs.com/docs/guide.html#indexes

MongoDB supports secondary indexes. With mongoose, we define these indexes within our Schema at the path level or the schema level. Defining indexes at the schema level is necessary when creating compound indexes.

  var animalSchema = new Schema({
    name: String,
    type: String,
    tags: { type: [String], index: true } // field level
  });

  animalSchema.index({ name: 1, type: -1 }); // schema level

When your application starts up, Mongoose automatically calls createIndex for each defined index in your schema. Mongoose will call createIndex for each index sequentially, and emit an 'index' event on the model when all the createIndex calls succeeded or when there was an error. While nice for development, it is recommended this behavior be disabled in production since index creation can cause a significant performance impact. Disable the behavior by setting the autoIndex option of your schema to false, or globally on the connection by setting the option autoIndex to false.

2.  MongoDB:  https://docs.mongodb.com/manual/indexes/#default-id-index
  MongoDB creates a unique index on the _id field during the creation of a collection. The _id index prevents clients from inserting two documents with the same value for the _id field. You cannot drop this index on the _id field.

3.  MongoDB createIndex():
https://docs.mongodb.com/manual/reference/method/db.collection.createIndex/#db.collection.createIndex