191. Logging out Users by sending invalid token
ref:  https://www.udemy.com/course/nodejs-express-mongodb-bootcamp/learn/lecture/15065666

Like signin user in , the process of signing is first to send a HTTP req but use a GET to get response with token will be sent back to browser.

1. In public\js\login.js, create function logout to be exported to index.js.

2.In public\js\index.js, import logout function from login.js. Then set a variable logOutBtn from '.nav__el--logout' element from _header.pug. Finally use a if statement when there's a logout button, set a event listener for logout function.

Then send a token which is not signed by JWT to client. So the the browser doesn't have signed token and will be considered not signed in (equals to signed out)

3. In controllers\authController.js, create logout function for sending a cookie with unsigned token (a plain string like "loggedOut")

4. In routes\userRoutes.js, use router.get('/logout', authController.logout); for the GET req in step 1. to handle the /logout URL.