200. Saving Image Name to Database #1 (save file name to document's photo field)
ref:  https://www.udemy.com/course/nodejs-express-mongodb-bootcamp/learn/lecture/15087352

Note: (ref:  https://github.com/expressjs/multer#usage  )
Multer adds a body object and a file or files object to the request object. The body object contains the values of the text fields of the form, the file or files object contains the files uploaded via the form.

So there will be a req.file obj after using uploadUserPhoto route handler function which uses multer package to upload file

1.In controllers\userController.js, use updatedUser function inside route handler function "updateMe" to update user's document with user id and object that has filtered data ( keys like name, email , photo in obj)