201. Turn uploaded image into Buffer and resizing Images with package 'sharp'
ref:  https://www.udemy.com/course/nodejs-express-mongodb-bootcamp/learn/lecture/15087354

1: In controllers\userController.js,
 The middleware handler function:  exports.uploadUserPhoto = upload.single('photo'); will upload the file to memory storage and save the file in req.file obj

Note #1:  The update Object's method "upload.single('photo')" will store the file in "req.file.buffer as an Buffer obj
ref:  https://github.com/expressjs/multer#singlefieldname

Note #2 In multer({ }) constructor function, when it takes argument obj as option has the "storage" option, the file in "req.file" will be turned into Buffer obj. And the Buffer obj will be in "req.file.buffer" for the next middleware to use or process.

2, The next middleware "resizeUserPhoto" will process (resize) the image with package "sharp" and its obj"sharp" constructor factory with parameters which are methods as different options

ref for parameters:  https://sharp.pixelplumbing.com/api-constructor